backend/requirements.txt
# Phase 3 Todo Chatbot Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
python-jose[cryptography]>=3.3.0
openai-agents>=0.0.3
mcp>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
ruff>=0.8.0
